Transition to angular representation of BRDF
the BRDF is now represented in data by each point corresponding to a vector. the vector describes the intensity in the direction the vector points. the length of the vector is the intensity.

Each Point now has three more attributes:
Azimuth: the azimuth angle of the vector, equal to the rotation of the vector around the z-axis. (-180,180]
Altitude: the altitude of the vector, the angle between the vector and the x,y-plane. [0,90]
length: the length of the vector, representing the intensity. [0,100]

changed sorting to sort based on azimuth, altitude, and length instead of coordinates.

removing overlap compares vectors in the same direction and keeps the longest.

rasterization normalized the angles to a given scale; CURRENTLY NOT COMPLETELY CORRECT!