added revolution of a list of points around an axis with demo data
added methodology to rotate a point around an axis using Rodrigues' rotation formula.
added methodology for rotating a number of point around an axis.
added operations for calculating the cross- and dot-product of two points, and multiplying the magnitude of a point.
added demonstration data for creating a BRDF using non-vertical axes